Fix compilation errors: remove duplicate Button.Style, fix InnerRadius property, and resolve ambiguous RelayCommand references
